cleanup: deleted unused files & functions renamed some files as well
