+++ seg_utils/ui/dialogs.py
@@ -1,133 +1,105 @@
 from PyQt5.QtWidgets import *
 from PyQt5.QtCore import QSize, QPoint, Qt
-from PyQt5.QtGui import QFont, QIcon
+from PyQt5.QtGui import QColor
 
-from seg_utils.ui.list_widgets import ListWidget
-from seg_utils.utils.qt import createListWidgetItemWithSquareIcon, get_icon
-
+from typing import List
 from pathlib import Path
 import os
 
-STYLESHEET = """QPushButton {
-                background-color: lightgray;
-                color: black;
-                min-height: 2em;
-                border-width: 2px;
-                border-radius: 8px;
-                border-color: black;
-                font: bold 12px;
-                padding: 2px;
-                %s
-                }
-                QPushButton::hover {
-                background-color: gray;
-                }
-                QPushButton::pressed {
-                border-style: outset;
-                }"""
-BUTTON_STYLESHEET = STYLESHEET % ""
-BUTTON_SELECTED_STYLESHEET = STYLESHEET % "border-style: outset;\nbackground-color: gray;"
+from seg_utils.ui.list_widgets import LabelList
+from seg_utils.utils.qt import get_icon
+from seg_utils.utils.stylesheets import BUTTON_STYLESHEET
 
 
-class NewLabelDialog(QDialog):
-    def __init__(self, parent: QWidget, *args):
-        super(NewLabelDialog, self).__init__(*args)
+class CloseMessageBox(QMessageBox):
+    def __init__(self, *args):
+        super(CloseMessageBox, self).__init__(*args)
 
-        # Default values
-        self.class_name = ""
-        self.parent = parent
+        quit_button = QPushButton(get_icon('quit'), "Quit Program")
+        cancel_button = QPushButton(self.style().standardIcon(QStyle.SP_DialogCancelButton), "Cancel")
+        self.addButton(quit_button, QMessageBox.AcceptRole)
+        self.addButton(cancel_button, QMessageBox.RejectRole)
 
-        # Create the shape of the QDialog
-        self.setFixedSize(QSize(300, 400))
-        move_to_center(self, self.parent.pos(), self.parent.size())
-        self.setWindowTitle("Select class of new shape")
+        self.setWindowTitle("Quit Program")
+        self.setText("Are you sure you want to quit?")
+        if self.parentWidget():
+            move_to_center(self, self.parentWidget().pos(), self.parentWidget().size())
 
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(10, 10, 10, 10)
-        layout.setSpacing(5)
 
-        # LineEdit
-        self.shapeSearch = QLineEdit(self)
-        font = QFont()
-        font.setPointSize(10)
-        font.setKerning(True)
-        self.shapeSearch.setFont(font)
-        self.shapeSearch.setPlaceholderText("Search shape label")
-        self.shapeSearch.setMaximumHeight(25)
-        self.shapeSearch.textChanged.connect(self.handle_shape_search)
+class CommentDialog(QDialog):
+    """QDialog to let the user enter notes regarding a specific annotation"""
 
-        # Buttons
-        button_widget = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        button_widget.accepted.connect(self.close)
-        button_widget.rejected.connect(self.on_cancel)
+    def __init__(self, comment: str):
+        super(CommentDialog, self).__init__()
+        self.setWindowTitle("Notes")
+        self.setFixedSize(500, 300)
+        self.comment = comment
 
-        # List of labels
-        self.listWidget = ListWidget(self)
-        self.listWidget.itemClicked.connect(self.on_list_selection)
+        # TextEdit where user can enter a comment
+        self.enter_comment = QTextEdit(self)
+        self.enter_comment.setText(self.comment)
 
-        # Combining everything
-        layout.addWidget(self.shapeSearch)
-        layout.addWidget(button_widget)
-        layout.addWidget(self.listWidget)
+        # Accept & cancel buttons
+        self.confirmation = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
+        self.confirmation.accepted.connect(self.create_comment)
+        self.confirmation.rejected.connect(self.close)
 
-        self.fill()
+        self.layout = QVBoxLayout(self)
+        self.layout.addWidget(self.enter_comment)
+        self.layout.addWidget(self.confirmation)
 
-    def fill(self):
-        """ get all the current label classes and display them in the dialog """
-        self.listWidget.clear()
-        for idx, _class in enumerate(self.parent.classes):
-            item = createListWidgetItemWithSquareIcon(_class, self.parent.colorMap[idx], 10)
-            self.listWidget.addItem(item)
-
-        # last entry in the list is used for creating a new label_class
-        item = QListWidgetItem(QIcon('icons/plus.png'), "New")
-        self.listWidget.addItem(item)
-
-    def handle_shape_search(self):
-        """ If user enters text, only suitable shapes are displayed """
-        text = self.shapeSearch.text()
-        for item_idx in range(self.listWidget.count()):
-            item = self.listWidget.item(item_idx)
-            if text not in item.text():
-                item.setHidden(True)
-            else:
-                item.setHidden(False)
-
-    def on_cancel(self):
-        self.class_name = ""
+    def create_comment(self):
+        """stores the written notes in the class variable and closes the dialog"""
+        self.comment = self.enter_comment.toPlainText()
         self.close()
 
-    def on_list_selection(self, item):
-        """ handles the functions to be executed when user selects something from the list"""
-        text = item.text()
 
-        # let user enter a new label class
-        if text == "New":
-            create_new_class = CreateNewClassDialog(self, self.parent.classes)
-            create_new_class.exec()
+class DeleteClassMessageBox(QMessageBox):
+    """ the answer variable can be interpreted as follows:
+        0 cancel
+        1 delete only the annotations of that class
+        2 delete the whole label class"""
+    def __init__(self, class_name: str, *args):
+        super(DeleteClassMessageBox, self).__init__(*args)
 
-            # if user entered a name, update the stored label classes
-            if create_new_class.new_class:
-                text = create_new_class.new_class
-                idx = len(self.parent.classes)
-                self.parent.classes[text] = idx
-                self.fill()
-                self.listWidget.item(self.listWidget.count() - 2).setSelected(True)  # highlight new item
+        self.setText("You are about to delete the '{}'-class.\nContinue?".format(class_name))
+        self.setInformativeText("This will remove all occurrences of '{}' from the image.\n\n".format(class_name))
+        self.setCheckBox(QCheckBox("Delete the label class from my list", self))
+        self.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
 
-            # otherwise, make sure that nothing is passed on
-            else:
-                text = ""
-                self.listWidget.currentItem().setSelected(False)
+        self.answer = 0
+        self.accepted.connect(self.set_answer)
 
-        # use selected/created label class to name the label
-        self.class_name = text
+    def set_answer(self):
+        """ sets the answer-variable """
+        if self.checkBox().isChecked():
+            self.answer = 2
+        else:
+            self.answer = 1
 
-    def set_text(self, text):
-        self.shapeSearch.setText(text)
-        # move the cursor to the end
-        new_cursor = self.shapeSearch.textCursor()
-        new_cursor.movePosition(self.shapeSearch.document().characterCount())
-        self.shapeSearch.setTextCursor(new_cursor)
+
+class DeleteFileMessageBox(QMessageBox):
+    def __init__(self, filename):
+        super(DeleteFileMessageBox, self).__init__()
+        self.setWindowTitle("Delete File")
+        self.setText("You are about to delete the file \n {}".format(filename))
+        self.setInformativeText("All annotations in this image will be lost.\n"
+                                "This operation can not be undone. \n Continue?")
+        self.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
+        self.setIcon(QMessageBox.Warning)
+
+
+class DeleteShapeMessageBox(QMessageBox):
+    def __init__(self, shape: str, *args):
+        super(DeleteShapeMessageBox, self).__init__(*args)
+        if self.parentWidget():
+            move_to_center(self, self.parentWidget().pos(), self.parentWidget().size())
+        reply = self.question(self, "Deleting Shape", f"You are about to delete {shape}.\nContinue?",
+                              QMessageBox.Yes | QMessageBox.No)
+        if reply == QMessageBox.Yes:
+            self.answer = 1
+        else:
+            self.answer = 0
 
 
 class ForgotToSaveMessageBox(QMessageBox):
@@ -146,174 +118,128 @@ class ForgotToSaveMessageBox(QMessageBox):
         self.addButton(cancel_button, QMessageBox.RejectRole)
         self.addButton(dismiss_button, QMessageBox.DestructiveRole)
 
-        move_to_center(self, self.parentWidget().pos(), self.parentWidget().size())
-        
-        
-class DeleteShapeMessageBox(QMessageBox):
-    def __init__(self, shape: str, *args):
-        super(DeleteShapeMessageBox, self).__init__(*args)
-        move_to_center(self, self.parentWidget().pos(), self.parentWidget().size())
-        reply = self.question(self, "Deleting Shape", f"You are about to delete {shape}. Continue?",
-                              QMessageBox.Yes | QMessageBox.No)
-        if reply == QMessageBox.Yes:
-            self.answer = 1
-        else:
-            self.answer = 0
+        if self.parentWidget():
+            move_to_center(self, self.parentWidget().pos(), self.parentWidget().size())
 
 
-class CloseMessageBox(QMessageBox):
-    def __init__(self, *args):
-        super(CloseMessageBox, self).__init__(*args)
+class SelectionDialog(QDialog):
+    """ a dialog that provides (a) a list with items where user can select from and search in
+        and (b) the possibility to create a new item from user input
 
-        quit_button = QPushButton(get_icon('quit'), "Quit Program")
-        cancel_button = QPushButton(self.style().standardIcon(QStyle.SP_DialogCancelButton), "Cancel")
-        self.addButton(quit_button, QMessageBox.AcceptRole)
-        self.addButton(cancel_button, QMessageBox.RejectRole)
+        used as scaffolding for other classes that may fill the list with items """
+    def __init__(self, selection_list, *args):
+        super(SelectionDialog, self).__init__(*args)
+        self.result = ""
 
-        self.setWindowTitle("Quit Program")
-        self.setText("Are you sure you want to quit?")
-        move_to_center(self, self.parentWidget().pos(), self.parentWidget().size())
+        self.setFixedSize(QSize(300, 400))
+        self.setLayout(QVBoxLayout())
+        self.layout().setContentsMargins(10, 10, 10, 10)
+        self.layout().setSpacing(5)
 
+        # Line Edit for searching/creating classes
+        self.input = QLineEdit()
+        self.input.setMaximumHeight(25)
 
-class DeleteClassMessageBox(QMessageBox):
-    def __init__(self, *args, class_name: str):
-        super(DeleteClassMessageBox, self).__init__(*args)
+        # info label for when user is about to create a new class
+        self.info = QLabel()
+        self.info.setContentsMargins(0, 0, 0, 0)
+        self.info.setFrameShape(QFrame.NoFrame)
+        self.info.setAlignment(Qt.AlignmentFlag.AlignCenter)
 
-        self.setText("You are about to delete the '{}'-class. Continue?".format(class_name))
-        self.setInformativeText("This will remove all occurrences of '{}' from the image.\n\n".format(class_name))
-        self.setCheckBox(QCheckBox("Do not delete the label class itself", self))
-        self.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
+        # list widget with the already existing classes
+        self.selection_list = selection_list
+        self.confirm = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
 
-        self.answer = 0
-        self.accepted.connect(self.set_answer)
+        # connect
+        self.input.textChanged.connect(self.handle_shape_input)
+        self.confirm.accepted.connect(self.on_close)
+        self.confirm.rejected.connect(self.on_cancel)
+        self.selection_list.itemClicked.connect(self.on_list_selection)
 
-    def set_answer(self):
-        """this sets the answer-variable.
-        0 cancel
-        1 delete only the annotations of that class
-        2 delete the whole class"""
-        if self.checkBox().isChecked():
-            self.answer = 1
-        else:
-            self.answer = 2
+        self.layout().addWidget(self.input)
+        self.layout().addWidget(self.info)
+        self.layout().addWidget(self.selection_list)
+        self.layout().addWidget(self.confirm)
 
+    def handle_shape_input(self):
+        """ function to filter the list according to user input """
+        self.info.clear()
+        self.result = ""
+        text = self.input.text()
+        matches = 0
 
-class CreateNewClassDialog(QDialog):
-    """ handles a QDialog used to let the user enter a new label class """
-    def __init__(self, parent, existing_classes: list, topic: str = "Label"):
-        super(CreateNewClassDialog, self).__init__()
+        # iterate through list, display only matches
+        for item_idx in range(self.selection_list.count()):
+            item = self.selection_list.item(item_idx)
+            item.setSelected(False)
+            if text not in item.text():
+                item.setHidden(True)
+            else:
+                item.setHidden(False)
+                matches += 1
 
-        self.setFixedSize(QSize(200, 120))
-        move_to_center(self, parent.pos(), parent.size())
-        self.setWindowTitle("Create new {} class".format(topic))
+        # option to create a new label class
+        if matches == 0 and text:
+            self.info.setStyleSheet("color: green;")
+            self.info.setText("Create new: {}".format(text))
+            self.result = text
 
-        # need it later to prevent duplicates
-        self.new_class = ""
-        self.existing_classes = existing_classes
+        # direct match: set selected
+        elif matches == 1:
+            for item_idx in range(self.selection_list.count()):
+                item = self.selection_list.item(item_idx)
+                if not item.isHidden():
+                    item.setSelected(True)
+                    self.on_list_selection(item)
 
-        # LineEdit
-        self.line_edit = QLineEdit(self)
-        font = QFont()
-        font.setPointSize(10)
-        font.setKerning(True)
-        self.line_edit.setFont(font)
-        self.line_edit.setPlaceholderText("Enter new {} name".format(topic))
-        self.line_edit.setMaximumHeight(25)
-
-        # Buttons
-        self.buttonWidget = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        self.buttonWidget.accepted.connect(self.ok_clicked)
-        self.buttonWidget.rejected.connect(self.close)
-
-        layout = QVBoxLayout(self)
-        layout.addWidget(self.line_edit)
-        layout.addWidget(self.buttonWidget)
-
-    def ok_clicked(self):
-        """ only accepts the new class if
-        (1) the name doesn't already exist and
-        (2) the name is not an empty string"""
-        name = self.line_edit.text()
-        if name in self.existing_classes:
-            self.line_edit.setText("")
-            self.line_edit.setPlaceholderText("'{}' already exists".format(name))
-        elif name:
-            self.new_class = name
-            self.close()
-
-
-class SelectPatientDialog(QDialog):
-    """
-    handles a QDialog to let the user select between video, image and whole slide image
-    the selected type is the used when importing a new file
-    """
-    def __init__(self, existing_patients: list):
-        super(SelectPatientDialog, self).__init__()
-
-        self.setFixedSize(QSize(300, 300))
-        self.setWindowTitle("Select Patient")
-        self.patient = ""
-
-        self.new_patient = QPushButton("New Patient")
-        self.new_patient.clicked.connect(self.create_new_patient)
-        self.new_patient.setStyleSheet(BUTTON_STYLESHEET)
-
-        self.confirmation = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        self.confirmation.button(QDialogButtonBox.Ok).setEnabled(False)
-        self.confirmation.rejected.connect(self.cancel)
-        self.confirmation.accepted.connect(self.close)
-
-        # text that will be displayed at the top
-        self.info = QLabel("Please specify a patient")
-        self.info.setStyleSheet("font: bold 12px")
-
-        # create patient_list list
-        self.patient_list = QListWidget()
-        self.patient_list.itemClicked.connect(self.set_patient)
-        self.patients = existing_patients
-        self.fill()
-
-        # set up the Dialog-Layout
-        self.layout = QVBoxLayout(self)
-        self.layout.addWidget(self.info)
-        self.layout.addWidget(self.patient_list)
-        self.layout.addWidget(self.new_patient)
-        self.layout.addWidget(self.confirmation)
-
-    def cancel(self):
-        """resets the class variable and closes the dialog"""
-        self.patient = ""
+    def on_cancel(self):
+        """ ensures that the result is an empty string"""
+        self.result = ""
         self.close()
 
-    def create_new_patient(self):
-        """lets user enter a new patient name/id"""
-        dlg = CreateNewClassDialog(self, self.patients, topic="Patient")
-        dlg.exec()
+    def on_close(self):
+        """ prevents an empty result when user clicks Ok"""
+        if self.result:
+            self.close()
+        else:
+            self.info.setStyleSheet("color: red;")
+            self.info.setText("Please select or create an item")
 
-        if dlg.new_class:
-            self.patients.append(dlg.new_class)
-            self.fill()
-            item = self.patient_list.item(self.patient_list.count() - 1)
-            item.setSelected(True)  # highlight new item
-            self.set_patient(item)
+    def on_list_selection(self, item: QListWidgetItem):
+        """ assigns the item name to the class variable"""
+        self.info.clear()
+        self.result = item.text()
 
-    def fill(self):
-        """fills the patient listWidget"""
-        self.patient_list.clear()
-        for patient in self.patients:
+
+class NewLabelDialog(SelectionDialog):
+    """ inherits the SelectionDialog, uses it to display label classes"""
+    def __init__(self, classes: List[str], color_map: List[QColor], *args):
+        super(NewLabelDialog, self).__init__(LabelList(), *args)
+        self.setWindowTitle("Select class of new shape")
+        self.input.setPlaceholderText("Enter shape label")
+
+        # fill list with the existing label classes
+        self.selection_list.update_with_classes(classes, color_map)
+
+
+class SelectPatientDialog(SelectionDialog):
+    """ inherits the SelectionDialog, uses it to display patient names"""
+    def __init__(self, existing_patients: list, *args):
+        super(SelectPatientDialog, self).__init__(QListWidget(), *args)
+        self.setWindowTitle("Select a patient")
+        self.input.setPlaceholderText("Enter patient name")
+
+        # fill list with existing patient names
+        for patient in existing_patients:
             item = QListWidgetItem(patient)
-            self.patient_list.addItem(item)
-
-    def set_patient(self, item):
-        """sets the patient class variable"""
-        self.patient = item.text()
-        self.confirmation.button(QDialogButtonBox.Ok).setEnabled(True)
+            self.selection_list.addItem(item)
 
 
 class ProjectHandlerDialog(QDialog):
-    def __init__(self, parent):
+    """ provides a dialog for the user to enter a project location and add initial files, if desired"""
+    def __init__(self):
         super(ProjectHandlerDialog, self).__init__()
-        self.setFixedSize(parent.size() / 2)
+        self.setFixedSize(800, 500)
         self.setWindowTitle('Create new Project')
 
         self.project_path = ""
@@ -388,11 +314,11 @@ class ProjectHandlerDialog(QDialog):
         """ function to let user select a file which will be added when the project is finally created"""
 
         # user first needs to specify the type of the file to be imported
-        select_patient = SelectPatientDialog(self.patients)
-        select_patient.exec()
-        self.patients = select_patient.patients
+        dlg = SelectPatientDialog(self.patients)
+        dlg.exec()
+        self.patients.append(dlg.result)
 
-        if select_patient.patient:
+        if dlg.result:
             filepath, _ = QFileDialog.getOpenFileName(self,
                                                       caption="Select File",
                                                       directory=str(Path.home()),
@@ -405,11 +331,11 @@ class ProjectHandlerDialog(QDialog):
                 msg.setIcon(QMessageBox.Information)
                 msg.setText("The file\n{}\nalready exists.\nOverwrite?".format(filename))
                 msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-                msg.accepted.connect(lambda: self.overwrite(filepath, filename, select_patient.patient))
+                msg.accepted.connect(lambda: self.overwrite(filepath, filename, dlg.result))
                 msg.exec()
             elif filename:
                 # TODO: Implement possibility to add several files at once
-                self.files[filepath] = select_patient.patient
+                self.files[filepath] = dlg.result
                 self.added_files.addItem(QListWidgetItem(filename))
 
     def check_path(self):
@@ -486,9 +412,8 @@ class ProjectHandlerDialog(QDialog):
             self.enter_path.setText(filename)
 
 
-
-
 def move_to_center(widget, parent_pos: QPoint, parent_size: QSize):
+    # TODO: implement move_to_center somewhere else, so the dialogs don't have to demand a parent widget
     r"""Moves the QDialog to the center of the parent Widget.
     As self.move moves the upper left corner to the place, one needs to subtract the own size of the window"""
     widget.move(parent_pos.x() + (parent_size.width() - widget.size().width())/2,
