+++ seg_utils/utils/stylesheets.py
@@ -1,14 +1,3 @@
-COMMENT_LIST = """QListWidget {
-                  color: rgb(0, 102, 204);
-                  selection-color: blue;
-                  selection-background-color: white;
-                  }
-                  
-                  QListWidget::item:hover {
-                  color: blue;
-                  } 
-                  """
-
 BUTTON_STYLESHEET = """QPushButton {
                        background-color: lightgray;
                        color: black;
